DEVELOPMENT_2026_08_15_1459, <COMPILE_SUCCESS> <RUNTIME_TEST>. PHASE: testing
implemented:
 - new development branch.
 - 8bit plaerId.
 - 64bit praiseId.

todo:
 - 8bit threadId in STRUCT concurrency.
 - pass threadId to std::thread on initialisation.

+++ CLIB_OpenEpiCentre/include/structs/CLIB_OpenEpiCentre_STRUCT__Output.h
@@ -32,8 +32,8 @@ namespace CLIBOpenEpiCentre
         static void stat_CLASS_boot4_INSTANTIATE_Output();
         static void stat_REG_boot0_DECLARE_Output();
     private:
-        static unsigned long* _REG_ptr_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId;
-        static unsigned long* _REG_ptr_Output_playerId;
+        static unsigned long long* _REG_ptr_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId;
+        static unsigned long long* _REG_ptr_Output_playerId;
         static std::list<Object*>* _REG_ptr_List_Of_PraiseOutputSubset;
         static void stat_REG_boot1_DEFINE_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId();
         static void stat_REG_boot1_DEFINE_Output_playerId();
@@ -44,8 +44,8 @@ namespace CLIBOpenEpiCentre
         static void stat_REG_boot3_INITIALISE_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId();
         static void stat_REG_boot3_INITIALISE_Output_playerId();
         static void stat_REG_boot3_INITIALISE_List_Of_PraiseOutputSubset(CLIB_OpenEpiCentre__Framework* obj, unsigned long long praiseId);
-        static unsigned long* stat_REG_get_ptr_Output_playerId();
-        static unsigned long* stat_REG_get_ptr_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId();
+        static unsigned long long* stat_REG_get_ptr_Output_playerId();
+        static unsigned long long* stat_REG_get_ptr_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId();
         static std::list<Object*>* stat_REG_get_ptr_Item_Of_ptr_Outputs_Subset();
     };
 }
